Additional airspeed sensor changes

+++ Autopilot/Path Manager/PathManager.c
@@ -517,7 +517,7 @@ void copyGPSData(){
         pmData.positionFix = (char)gpsData.positionFix;
     }
     pmData.batteryLevel = getCurrentPercent();
-    pmData.airspeed = (float)getCurrentAirspeed();
+    pmData.airspeed = getCurrentAirspeed();
     pmData.altitude = getAltitude(); //gpsData.altitude; //want to get altitude regardless of if there is new GPS data
     pmData.checkbyteDMA = generatePMDataDMAChecksum();
 }
